backend/requirements.txt
asgiref==3.6.0
cffi==1.15.1
cryptography==38.0.4
Django==4.1.4
django-cors-headers==3.13.0
django-environ==0.9.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
jwt==1.3.1
pycparser==2.21
pytz==2022.7
simplejwt==2.0.1
sqlparse==0.4.3
typing==3.7.4.3
